Rename interface and constant file of icons folder
