+++ src/components/UI/Icons/Icons.constants.ts
@@ -6,7 +6,7 @@ import {
   AccountBalance, MedicalInformation, FamilyRestroom, ShoppingCart,
   Menu,
 } from '@mui/icons-material';
-import { AllCategoryIcons, AppIcons } from './interface';
+import { AllCategoryIcons, AppIcons } from './Icons.interface';
 import { AppColors } from '../../../styles';
 
 export const APP_ICONS: AppIcons = {
