+++ src/components/UI/Records/features/RecordDrawer/RecordDrawer.tsx
@@ -4,7 +4,7 @@ import {
 import { useNavigate } from 'react-router-dom';
 
 import { RecordDrawerProps } from '../../interface';
-import { AllCategoryIcons } from '../../../Icons/interface';
+import { AllCategoryIcons } from '../../../Icons/Icons.interface';
 import { EDIT_RECORD_ROUTE } from '../../../../../pages/RoutesConstants';
 import { setRecordToBeModified } from '../../../../../redux/slices/Records/records.slice';
 import { useAppDispatch, useAppSelector } from '../../../../../redux/hooks';
